Create Easy-1.c
This code works by iterating through the string from the end and keeping track of the length of the current word. When a space is encountered, the length of the current word is updated and the flag indicating that a word is being counted is reset. The final length of the last word is returned.